Remove the dist file after moving bundled files into add-on assets
