+++ add-ons/bp-search-block/assets/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('bp-block-data', 'lodash', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-element', 'wp-i18n'), 'version' => 'd72e444252b5b760d333');
