+++ add-ons/bp-search-block/assets/view.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('wp-dom-ready'), 'version' => '3516006aefaea2403f1a');
