+++ add-ons/bp-search-block/assets/view.js
