--- add-ons/bp-search-block/block.json
@@ -1,55 +1,0 @@
-{
-	"apiVersion": "2",
-	"name": "bp/search-form",
-	"title": "Community Search",
-	"category": "buddypress",
-	"icon": "search",
-	"description": "A Block to search for posts, sites, activities, members or groups from any post, page or widget of your BuddyPress powered community site!",
-	"keywords": [ "BuddyPress", "search", "community" ],
-	"textdomain": "bp-search-block",
-	"attributes": {
-		"label": {
-			"type": "string",
-			"source": "html",
-			"selector": ".bp-search-label",
-			"default": "Search"
-		},
-		"useLabel": {
-			"type": "boolean",
-			"default": true
-		},
-		"buttonText": {
-			"type": "string",
-			"source": "html",
-			"selector": ".bp-search-button",
-			"default": "Search"
-		},
-		"useIcon": {
-			"type": "boolean",
-			"default": false
-		},
-		"placeholder": {
-			"type": "string",
-			"default": ""
-		},
-		"activeOptions": {
-			"type": "array",
-			"default": ["posts"]
-		},
-		"defaultOption": {
-			"type": "string",
-			"default": "posts"
-		},
-		"action": {
-			"type": "string",
-			"default": ""
-		}
-	},
-	"supports": {
-		"align": true
-	},
-	"editorScript": "file:js/index.js",
-	"script": "file:js/view.js",
-	"editorStyle": "file:css/index.css",
-	"style": "file:css/style.css"
-}
