--- add-ons/bp-search-block/css/index.css
@@ -1,2 +1,0 @@
-.wp-block[data-align=center] .wp-block-bp-search-form .bp-block-search__inside-wrapper{margin:auto}.wp-block-bp-search-form .bp-search-button{height:auto;border-radius:initial;display:flex;align-items:center}.wp-block-bp-search-form__components-button-group{margin-top:10px}
-/*# sourceMappingURL=/index.css.map */
