--- add-ons/bp-search-block/css/style.css
@@ -1,2 +1,0 @@
-.bp-search-button{background:#f7f7f7;border:1px solid #ccc;padding:.375em .625em;color:#32373c;margin-left:.625em;word-break:normal;font-size:inherit;font-family:inherit;line-height:inherit}.bp-search-button.has-icon{line-height:0}.bp-block-search__inside-wrapper{display:flex;flex:auto;flex-wrap:nowrap;max-width:100%}.bp-block-search-for__label,.bp-search-label{width:100%}.bp-block-search__input{padding:8px;flex-grow:1;min-width:3em;border:1px solid #949494;font-size:inherit;font-family:inherit;line-height:inherit}.bp-block-search.bp-block-search__button-only .bp-search-button{margin-left:0}.bp-block-search.bp-block-search__button-inside .bp-block-search__inside-wrapper{padding:4px;border:1px solid #949494}.bp-block-search.bp-block-search__button-inside .bp-block-search__inside-wrapper .bp-block-search__input{border-radius:0;border:none;padding:0 0 0 .25em}.bp-block-search.bp-block-search__button-inside .bp-block-search__inside-wrapper .bp-block-search__input:focus{outline:none}.bp-block-search.bp-block-search__button-inside .bp-block-search__inside-wrapper .bp-search-button{padding:.125em .5em}.bp-block-search.aligncenter .bp-block-search__inside-wrapper{margin:auto}.bp-block-search-for__wrapper{margin:.5em 0}.bp-block-search-for__wrapper .bp-block-search-for__options{list-style:none;padding:0;margin:.2em 0}.bp-block-search-for__wrapper .bp-block-search-for__options li label input[type=radio]{display:inline-block;margin-right:.5em}
-/*# sourceMappingURL=/style.css.map */
