--- add-ons/bp-search-block/js/index.asset.php
@@ -1,1 +1,0 @@
-<?php return array('dependencies' => array('lodash', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-element', 'wp-i18n', 'bp-block-data'), 'version' => 'eeb1f33dfffdab970a1c');
