--- add-ons/bp-search-block/js/view.asset.php
@@ -1,1 +1,0 @@
-<?php return array('dependencies' => array('wp-dom-ready'), 'version' => '43e5b59c875857271585');
