+++ src/add-ons/bp-search-block/editor-style.scss
