+++ src/add-ons/bp-search-block/imports/edit.js
