+++ src/add-ons/bp-search-block/imports/option-checkbox.js
