+++ src/add-ons/bp-search-block/view-style.scss
